+++ demo/admin/src/products/generator/ProductsPage.tsx
@@ -63,10 +63,13 @@ export function ProductsPage() {
                                 </IconButton>
                             )}
                             actionsColumnWidth={116}
+                            onRowClick={(params) => {
+                                productsStackSwitchApi.activatePage("edit", params.row.id);
+                            }}
                         />
                     </StackMainContent>
                 </StackPage>
-                <StackPage name="edit" title={intl.formatMessage({ id: "products.editProduct", defaultMessage: "Edit Product" })}>
+                <StackPage name="edit" title={intl.formatMessage({ id: "products.editProduct", defaultMessage: "Edit product" })}>
                     {(selectedProductId) => (
                         <SaveBoundary>
                             <>
@@ -158,7 +161,7 @@ export function ProductsPage() {
                         </SaveBoundary>
                     )}
                 </StackPage>
-                <StackPage name="add" title={intl.formatMessage({ id: "products.addProduct", defaultMessage: "Add Product" })}>
+                <StackPage name="add" title={intl.formatMessage({ id: "products.addProduct", defaultMessage: "Add product" })}>
                     <SaveBoundary>
                         <FormToolbar />
                         <MainContent>