+++ packages/admin/admin-generator/src/commands/generate/generate-command.ts
@@ -53,7 +53,7 @@ function isComponentFormFieldConfig(arg: any): arg is ComponentFormFieldConfig {
     return arg && arg.type === "component";
 }
 
-export type StaticSelectValue = { value: string; label: string } | string;
+export type StaticSelectValue = { value: string; label: string | FormattedMessageElement } | string;
 type AsyncSelectFilter =
     | {
           /**
@@ -99,7 +99,7 @@ export type FormFieldConfig<T> = (
           disableSlider?: boolean;
           initialValue?: { min: number; max: number };
       } & InputBaseFieldConfig)
-    | { type: "boolean"; name: UsableFormFields<T>; initialValue?: boolean }
+    | { type: "boolean"; name: UsableFormFields<T>; initialValue?: boolean; checkboxLabel?: string | FormattedMessageElement }
     | ({ type: "date"; name: UsableFormFields<T>; initialValue?: string } & InputBaseFieldConfig)
     | ({ type: "dateTime"; name: UsableFormFields<T>; initialValue?: Date } & InputBaseFieldConfig)
     | ({