+++ packages/api/api-generator/src/commands/generate/utils/__tests__/find-hooks-service.spec.ts
@@ -1,15 +1,22 @@
-import { CrudGenerator, CrudGeneratorHooksService, CurrentUser } from "@comet/cms-api";
+import { CrudGenerator, CrudGeneratorHooksService, CurrentUser, MutationError } from "@comet/cms-api";
 import { BaseEntity, defineConfig, Entity, MikroORM, PrimaryKey, Property } from "@mikro-orm/postgresql";
 import { LazyMetadataStorage } from "@nestjs/graphql/dist/schema-builder/storages/lazy-metadata.storage";
 import { v4 as uuid } from "uuid";
 
 import { findHooksService } from "../find-hooks-service";
 import { testPermission } from "../test-helper";
+import { TestEntity } from "./find-hooks-service/test-entity.entity";
+import { TestEntityService } from "./find-hooks-service/test-entity.service";
+import { TestEntity3 } from "./find-hooks-service/test-entity3.entity";
 
-class TestEntityService implements CrudGeneratorHooksService {
+export class Test2MutationError implements MutationError {
+    code: string;
+}
+
+class TestEntity2Service implements CrudGeneratorHooksService {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    async validateCreateInput(input: any, options: { currentUser: CurrentUser; args: { xxx: string } }): Promise<void> {
-        // no-op
+    async validateCreateInput(input: any, options: { currentUser: CurrentUser; args: { xxx: string } }): Promise<Test2MutationError[]> {
+        return [];
     }
 }
 
@@ -17,9 +24,9 @@ class TestEntityService implements CrudGeneratorHooksService {
 @CrudGenerator({
     targetDirectory: __dirname,
     requiredPermission: testPermission,
-    hooksService: TestEntityService,
+    hooksService: TestEntity2Service,
 })
-export class TestEntity extends BaseEntity {
+export class TestEntity2 extends BaseEntity {
     @PrimaryKey({ type: "uuid" })
     id: string = uuid();
 
@@ -29,6 +36,36 @@ export class TestEntity extends BaseEntity {
 
 describe("find-hooks-service", () => {
     it("finds hooks service and defined methods", async () => {
+        LazyMetadataStorage.load();
+        const orm = await MikroORM.init(
+            defineConfig({
+                dbName: "test-db",
+                connect: false,
+                entities: [TestEntity2],
+            }),
+        );
+
+        const hooksService = findHooksService({
+            generatorOptions: { targetDirectory: __dirname, requiredPermission: testPermission, hooksService: TestEntityService },
+            metadata: orm.em.getMetadata().get("TestEntity2"),
+        });
+        if (!hooksService) throw new Error("hooksService not found");
+        expect(hooksService.className).toEqual("TestEntity2Service");
+        expect(hooksService.imports).toEqual([
+            {
+                name: "Test2MutationError",
+                importPath: "./find-hooks-service.spec",
+            },
+        ]);
+        expect(hooksService.validateCreateInput).toBeDefined();
+        expect(hooksService.validateCreateInput?.options).toContain("currentUser");
+        expect(hooksService.validateCreateInput?.options).toContain("args");
+        expect(hooksService.validateCreateInput?.returnType).toEqual("Test2MutationError");
+        expect(hooksService.validateUpdateInput).toBeNull();
+
+        await orm.close();
+    });
+    it("finds hooks service and imports with error in same file as service", async () => {
         LazyMetadataStorage.load();
         const orm = await MikroORM.init(
             defineConfig({
@@ -44,10 +81,51 @@ describe("find-hooks-service", () => {
         });
         if (!hooksService) throw new Error("hooksService not found");
         expect(hooksService.className).toEqual("TestEntityService");
-        expect(hooksService.imports).toEqual([]);
+        expect(hooksService.imports).toEqual([
+            {
+                name: "TestEntityService",
+                importPath: "./find-hooks-service/test-entity.service",
+            },
+            {
+                name: "TestMutationError",
+                importPath: "./find-hooks-service/test-entity.service",
+            },
+        ]);
         expect(hooksService.validateCreateInput).toBeDefined();
-        expect(hooksService.validateCreateInput?.options).toContain("currentUser");
-        expect(hooksService.validateCreateInput?.options).toContain("args");
+        expect(hooksService.validateCreateInput?.returnType).toEqual("TestMutationError");
+        expect(hooksService.validateUpdateInput).toBeNull();
+
+        await orm.close();
+    });
+
+    it("finds hooks service and imports with error in other file as service", async () => {
+        LazyMetadataStorage.load();
+        const orm = await MikroORM.init(
+            defineConfig({
+                dbName: "test-db",
+                connect: false,
+                entities: [TestEntity3],
+            }),
+        );
+
+        const hooksService = findHooksService({
+            generatorOptions: { targetDirectory: __dirname, requiredPermission: testPermission, hooksService: TestEntityService },
+            metadata: orm.em.getMetadata().get("TestEntity3"),
+        });
+        if (!hooksService) throw new Error("hooksService not found");
+        expect(hooksService.className).toEqual("TestEntity3Service");
+        expect(hooksService.imports).toEqual([
+            {
+                name: "TestEntity3Service",
+                importPath: "./find-hooks-service/test-entity3.service",
+            },
+            {
+                name: "TestEntity3MutationError",
+                importPath: "./find-hooks-service/test-entity3-error",
+            },
+        ]);
+        expect(hooksService.validateCreateInput).toBeDefined();
+        expect(hooksService.validateCreateInput?.returnType).toEqual("TestEntity3MutationError");
         expect(hooksService.validateUpdateInput).toBeNull();
 
         await orm.close();