+++ packages/api/api-generator/src/commands/generate/utils/__tests__/find-hooks-service/test-entity.entity.ts
@@ -1,0 +1,20 @@
+import { CrudGenerator } from "@comet/cms-api";
+import { BaseEntity, Entity, PrimaryKey, Property } from "@mikro-orm/postgresql";
+import { v4 as uuid } from "uuid";
+
+import { testPermission } from "../../test-helper";
+import { TestEntityService } from "./test-entity.service";
+
+@Entity()
+@CrudGenerator({
+    targetDirectory: __dirname,
+    requiredPermission: testPermission,
+    hooksService: TestEntityService,
+})
+export class TestEntity extends BaseEntity {
+    @PrimaryKey({ type: "uuid" })
+    id: string = uuid();
+
+    @Property()
+    foo: string;
+}