+++ packages/admin/admin-rte/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @comet/admin-rte
 
+## 8.16.0
+
+### Patch Changes
+
+- Updated dependencies [472b496]
+    - @comet/admin@8.16.0
+    - @comet/admin-icons@8.16.0
+
+## 8.15.0
+
+### Patch Changes
+
+- Updated dependencies [cc96333]
+- Updated dependencies [180d1e3]
+- Updated dependencies [f7b9de9]
+- Updated dependencies [39a9bb0]
+- Updated dependencies [5b52998]
+- Updated dependencies [0c2435a]
+    - @comet/admin@8.15.0
+    - @comet/admin-icons@8.15.0
+
 ## 8.14.0
 
 ### Patch Changes