+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/eslint-config
 
+## 8.16.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.16.0
+
+## 8.15.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.15.0
+
 ## 8.14.0
 
 ### Minor Changes