+++ packages/mail-react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @comet/mail-react
 
+## 8.16.0
+
+## 8.15.0
+
 ## 8.14.0
 
 ### Minor Changes